Update flmd_quick_guide.md
I added the text below rather than completely referring the user back to another reporting format. There could/should be a cross-link to the "Delimiter section of the CSV Reporting Format Detailed Guide"

"If entering more than one missing value code, use a vertical bar "|" or semicolon ";" instead of a comma or protect the comma with matching double quotation marks around the entire value. For more information about commas not meant to be a delimiter (e.g. used within a cell), refer to the Delimiter section of the CSV Reporting Format Detailed Guide."